bugfix0.096

+++ ais-bench_workload/tool/find_best_batchsize/test/test_aoe_mode.sh
@@ -21,7 +21,6 @@ cmd="bash ../find_best_batchsize.sh
     --job_type 1
     "
 [ $cmd != $ret_ok ] && { echo "command failed";return 1; }
-#$cmd
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -37,8 +36,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 0
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -55,8 +53,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -74,8 +71,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
